Initial Config for Puppeteer
Base config for puppeteer and replace PhantomJS
